+++ edgaro/explain/explainer.py
@@ -10,7 +10,7 @@ from typing import List, Optional, Literal
 
 from edgaro.data.dataset import Dataset
 from edgaro.model.model import Model
-from edgaro.explain.explainer_result import ModelProfileExplanation, Curve
+from edgaro.explain.explainer_result import Explanation, ModelProfileExplanation, Curve, ModelPartsExplanation
 from edgaro.base.utils import print_unbuffered
 
 
@@ -32,10 +32,18 @@ class Explainer:
     N : int, optional, default=None
         Number of observations that will be sampled from the test Dataset before the calculation of profiles
         (PDP/ALE curves). None means all data.
-    curve_type : {'PDP', 'ALE'}, default='PDP'
-        A curve type to be calculated.
+    explanation_type : {'PDP', 'ALE', 'VI'}, default='PDP'
+        An explanation type to be calculated (`PDP` - Partial Dependence Profile, `ALE` - Accumulated Local Effects,
+        `VI` - Variable Importance)
     verbose : bool, default=False
         Print messages during calculations.
+    processes : int, default=1
+        Number of processes for the calculation of explanations.
+        If -1, it is replaced with the number of available CPU cores.
+    random_state : int, optional, default=None
+        Random state seed.
+    B : int, optional, default=10
+        Number of permutation rounds to perform on each variable - applicable only if explanation_type='VI'.
 
     Attributes
     ----------
@@ -46,15 +54,19 @@ class Explainer:
     N : int, optional, default=None
         Number of observations that will be sampled from the test Dataset before the calculation of profiles
         (PDP/ALE curves). None means all data.
-    curve_type : {'PDP', 'ALE'}
-        A curve type to be calculated.
+    explanation_type : {'PDP', 'ALE', 'VI'}
+        An explanation type to be calculated.
     verbose : bool
         Print messages during calculations.
     explainer : dx.Explainer, optional
         An explainer object from `dalex` package.
-    processes : int, default=1
+    processes : int
         Number of processes for the calculation of explanations.
         If -1, it is replaced with the number of available CPU cores.
+    random_state : int, optional
+        Random state seed
+    B : int, optional
+        Number of permutation rounds to perform on each variable - applicable only if explanation_type='VI'.
 
     References
     ----------
@@ -63,15 +75,18 @@ class Explainer:
 
     """
 
-    def __init__(self, model: Model, N: Optional[int] = None, curve_type: Literal['PDP', 'ALE'] = 'PDP',
-                 verbose: bool = False, processes: int = 1) -> None:
+    def __init__(self, model: Model, N: Optional[int] = None, explanation_type: Literal['PDP', 'ALE', 'VI'] = 'PDP',
+                 verbose: bool = False, processes: int = 1, random_state: Optional[int] = None,
+                 B: Optional[int] = 10) -> None:
         self.model = model
         self.explainer = None
         self.name = model.name
         self.N = N
-        self.curve_type = curve_type
+        self.explanation_type = explanation_type
         self.verbose = verbose
         self.processes = processes
+        self.random_state = random_state
+        self.B = B
 
         if self.processes == -1:
             self.processes = multiprocessing.cpu_count()
@@ -105,18 +120,18 @@ class Explainer:
         if self.verbose:
             print_unbuffered(f'dalex explainer inside {self.__repr__()} was created with {dataset.name}')
 
-    def transform(self, variables: Optional[List[str]] = None) -> ModelProfileExplanation:
+    def transform(self, variables: Optional[List[str]] = None) -> Explanation:
         """
-        Calculate the curve.
+        Calculate the explanation.
 
         Parameters
         ----------
         variables : list[str], optional
-            List of variables for which the curves should be calculated.
+            List of variables for which the explanation should be calculated.
 
         Returns
         -------
-        ModelProfileExplanation
+        Explanation
         """
         if self.explainer is None:
             raise Exception('Explainer was not fitted!')
@@ -126,34 +141,53 @@ class Explainer:
         if variables is None:
             variables = list(self.model.get_train_dataset().data.columns)
 
-        if self.curve_type == 'PDP':
-            curve_type = 'partial'
-        elif self.curve_type == 'ALE':
-            curve_type = 'accumulated'
+        if self.verbose:
+            print_unbuffered(f'{self.explanation_type} is being calculated in {self.__repr__()} for '
+                             f'{self.model.get_test_dataset().name}')
+
+        if self.explanation_type == 'PDP' or self.explanation_type == 'ALE':
+            explanation_type = 'partial' if self.explanation_type == 'PDP' else 'accumulated'
+
+            self.__transform_curve_category(dict_output, variables, category_colnames_base, explanation_type)
+            self.__transform_curve_other(dict_output, variables, category_colnames_base, explanation_type)
+
+            if self.verbose:
+                print_unbuffered(f'{self.explanation_type} was calculated calculated in {self.__repr__()} for '
+                                 f'{self.model.get_test_dataset().name}')
+
+            return ModelProfileExplanation(dict_output, self.name, self.model.get_category_colnames())
+        elif self.explanation_type == 'VI':
+            explanation_type = 'variable_importance'
+            self.__transform_feature_importance(dict_output, variables, explanation_type)
+
+            if self.verbose:
+                print_unbuffered(f'{self.explanation_type} was calculated calculated in {self.__repr__()} for '
+                                 f'{self.model.get_test_dataset().name}')
+
+            return ModelPartsExplanation(dict_output, self.name)
         else:
             raise Exception('Wrong curve type!')
 
-        if self.verbose:
-            print_unbuffered(f'{self.curve_type} is being calculated in {self.__repr__()} for '
-                             f'{self.model.get_test_dataset().name}')
-
+    def __transform_curve_category(self, dict_output, variables, category_colnames_base, explanation_type):
         category_colnames = list(set(variables).intersection(set(category_colnames_base)))
         if len(category_colnames) > 0:
-            for col in category_colnames:
-                out_category = self.explainer.model_profile(verbose=False, variables=[col],
-                                                            variable_type='categorical',
-                                                            N=self.N, type=curve_type,
-                                                            processes=self.processes)
-                y = out_category.result['_yhat_']
-                x = out_category.result['_x_']
-                dict_output[str(col)] = Curve(x, y)
+            out_category = self.explainer.model_profile(verbose=False, variables=category_colnames,
+                                                        variable_type='categorical',
+                                                        N=self.N, type=explanation_type,
+                                                        processes=self.processes, random_state=self.random_state)
+            y = out_category.result['_yhat_']
+            x = out_category.result['_x_']
+            variable_names = out_category.result['_vname_'].unique()
+            for i in range(len(variable_names)):
+                dict_output[str(variable_names[i])] = Curve(x, y)
 
+    def __transform_curve_other(self, dict_output, variables, category_colnames_base, explanation_type):
         other_colnames = list(set(variables).difference(set(category_colnames_base)))
         if len(other_colnames) > 0:
             out_others = self.explainer.model_profile(verbose=False, variables=other_colnames,
                                                       variable_type='numerical',
-                                                      N=self.N, type=curve_type,
-                                                      processes=self.processes)
+                                                      N=self.N, type=explanation_type,
+                                                      processes=self.processes, random_state=self.random_state)
             variable_names = out_others.result['_vname_'].unique()
             y = out_others.result['_yhat_']
             x = out_others.result['_x_']
@@ -163,14 +197,19 @@ class Explainer:
                 higher = int((i + 1) * length)
                 dict_output[str(variable_names[i])] = Curve(x[lower:higher], y[lower:higher])
 
-        if self.verbose:
-            print_unbuffered(f'{self.curve_type} was calculated calculated in {self.__repr__()} for '
-                             f'{self.model.get_test_dataset().name}')
+    def __transform_feature_importance(self, dict_output, variables, explanation_type):
+        out = self.explainer.model_parts(verbose=False, variables=variables,
+                                         N=self.N, B=self.B, type=explanation_type,
+                                         processes=self.processes, random_state=self.random_state)
+        out = out.result
+        out = out[(out.variable != '_full_model_') & (out.variable != '_baseline_')]
+        out = out[['variable', 'dropout_loss']]
 
-        return ModelProfileExplanation(dict_output, self.name, self.model.get_category_colnames())
+        for index, row in out.iterrows():
+            dict_output[row[0]] = row[1]
 
     def __str__(self) -> str:
-        return f"Explainer for model {self.name} with {self.curve_type} curve type"
+        return f"Explainer for model {self.name} with {self.explanation_type}  explanation type"
 
     def __repr__(self) -> str:
-        return f"<Explainer for model {self.name} with {self.curve_type} curve type>"
+        return f"<Explainer for model {self.name} with {self.explanation_type} explanation type>"